refactor context
1. remove serdes related code to federation
2. add destroy api for computing and federation
