+++ python/fate/arch/context/__init__.py
@@ -15,6 +15,6 @@
 from ._cipher import CipherKit, PHECipher, PHECipherPublic
 from ._context import Context
 from ._namespace import NS
-from ._federation import Parties
+from ._parties import Parties
 
 __all__ = ["Context", "CipherKit", "PHECipher", "PHECipherPublic", "NS", "Parties"]
