+++ python/fate/arch/federation/api/__init__.py
@@ -1,3 +1,13 @@
 from ._federation import Federation
-from ._type import FederationDataType, PartyMeta
+from ._serdes import TableRemotePersistentPickler, TableRemotePersistentUnpickler
 from ._table_meta import TableMeta
+from ._type import FederationDataType, PartyMeta
+
+__all__ = [
+    "Federation",
+    "TableMeta",
+    "PartyMeta",
+    "FederationDataType",
+    "TableRemotePersistentPickler",
+    "TableRemotePersistentUnpickler",
+]
