+++ python/fate/arch/federation/backends/pulsar/_federation.py
@@ -147,7 +147,7 @@ class PulsarFederation(MessageQueueBasedFederation):
     def __getstate__(self):
         pass
 
-    def destroy(self):
+    def _destroy(self):
         # The idea cleanup strategy is to consume all message in topics,
         # and let pulsar cluster to collect the used topics.
